pyproject.toml
[project]
name = "sk-handofforchestration"
version = "0.1.0"
description = "Semantic Kernel Agent Handoff Orchestration with Image Analysis"
authors = [
    {name = "Yu-Hong Lin", email = "yu-hong.lin@microsoft.com"}
]
requires-python = ">=3.10"
dependencies = [
    "semantic-kernel>=1.14.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = []
